chart stuff + optimizations + more

+++ src/spodcat/serializers/charts.py
@@ -4,7 +4,7 @@ from rest_framework import serializers
 # pylint: disable=abstract-method
 class ChartDataPointSerializer(serializers.Serializer):
     x = serializers.IntegerField()
-    y = serializers.IntegerField()
+    y = serializers.FloatField()
 
 
 # pylint: disable=abstract-method
